After many many... many hours fixed SSL certificate bug rails has on windows 10
This fix has been added and should mean no one else has to deal with it.

But for reference this is the fix.
Download cacerts.pem to some directory pref same as ruby script you are firing
http://curl.haxx.se/ca/cacert.pem

Add the following to your script:
    # Based on fix described at: https://gist.github.com/fnichol/867550
    ENV['SSL_CERT_FILE'] = File.expand_path(File.dirname(__FILE__)) + "/config/cacert.pem"

Started working on getting a base setup with Octokit